refactor: extract DownloadsSettingsFragment business logic into UseCase
- Extract rebuildDownloadsIndex(), matchChapters(), extractChapterNumber(),
  getNormalizedSourceName(), isSameOrSimilarSource(), showMigrationDialog()
  into DownloadsIndexRebuildUseCase
- Fragment reduced from 648 to 394 lines
- All logic is VERBATIM — no behavior changes
- UseCase is @Singleton, Hilt-injectable, testable in isolation
- Fragment delegates to UseCase via callback pattern for migration dialog

Note: AppSettings.kt split (1,219 lines) requires more careful planning
due to its 100+ import sites and interdependent properties. Will be
addressed in a follow-up PR.
